deleted userContext.js and moved context into APP.js to also include username and loggedIn vars
