+++ src/client/App.js
@@ -9,11 +9,11 @@ import LoginModal from './components/LoginModal';
 import RegistrationModal from './components/RegistrationModal';
 import Profile from './pages/Profile';
 import Navbar from './components/Navbar';
-import UserContext from './userContext';
 
 export default function App() {
   const [loggedIn, setLoggedIn] = React.useState(false);
   const [username, changeUsername] = React.useState('');
+  const [token, changeToken] = React.useState('');
 
   function loginUser() {
     setLoggedIn(true);
@@ -23,10 +23,33 @@ export default function App() {
     changeUsername(user);
   }
 
+  // User Context
+  const UserContext = React.createContext({
+    username: { username },
+    loggedIn: { loggedIn },
+    token: { token },
+    changeToken: { changeToken }
+    // ^ dependent on our implementation and whether or not we will need this method
+    // This is automatically passed with the context.
+
+    // Will be providing in App.js by wrapping child components with a provider:
+    /* <UserContext.Provider value={token, changeToken}>
+            <Layout />
+          </UserContext.Provider> */
+
+    // on value change of 'token', child props will be rerendered
+
+    // For context consumption:
+    /* <UserContext.Consumer>
+              {token => (
+                <childElement token = {token} />
+              )}
+            </UserContext.Consumer> */
+  });
 
   return (
     <>
-      <UserContext.Provider value={UserContext.token}>
+      <UserContext.Provider value={token}>
         <Navbar loggedIn={loggedIn} username={username} />
         <Switch>
           <Route exact path="/" component={Home} />
